rendering precomputed scores
