+++ configs/environment.yaml
@@ -1,0 +1,12 @@
+name: ditto-ui
+
+channels:
+  - conda-forge
+  - anaconda
+  - bioconda
+
+dependencies:
+  - python=3.10.11
+  - pandas=2.0.1
+  - bcftools=1.21
+  - streamlit=1.39.0
