--- requirements.txt
@@ -1,12 +1,0 @@
-tensorflow==2.12
-streamlit==1.28.2
-pandas==2.1.1
-shap==0.42.1
-protobuf==3.20.3
-#python==3.11.5
-matplotlib==3.8.0
-scikit-learn==1.3.0
-scipy==1.11.3
-numpy==1.23.5
-pyaml==23.9
-plotly==5.18
